Move VideoReader out of __init__
